+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add Streaming Support for WebSockets and Grafana Live (#113)
 - Move InfluxDB Example to the Guest Blog post (#115)
 - Update Documentation and Tutorials in README (#116)
+- Add Data Zoom to default Line Example (#117)
 
 ## 4.0.0 (2022-12-20)